Move WebsiteM and RequestM to the Types module
