+++ pf-backend/src/Website/API/Blog.hs
@@ -23,7 +23,7 @@ import Website.API.Auth
 import Website.API.Common
 import Website.Models
 import Website.Utils
-import Website.WebsiteM
+import Website.Types
 
 
 type BlogPostAPI =
